-Change Service to pull all items rather than just selected items

+++ src/main/java/com/srobinson24/sandbox/tradeskillmaster/processor/ProfitProcessor.java
@@ -6,9 +6,9 @@ import com.srobinson24.sandbox.tradeskillmaster.domain.Enchant;
  * Created by srobinso on 4/10/2017.
  */
 public interface ProfitProcessor {
-    int  getCraftingCost(Enchant enchant);
+    long getCraftingCost(Enchant enchant);
 
-    int calculateProfit(Enchant enchant);
+    long calculateProfit(Enchant enchant);
 
-    int truncateSilverAndCopper(int value);
+    long truncateSilverAndCopper(long value);
 }
